requirements.txt
aiocoap==0.4.7
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.6
colorama==0.4.6
confluent-kafka==2.2.0
docopt==0.6.2
filelock==3.9.0
Flask==2.3.2
flask-mysqldb==0.2.0
Flask-SQLAlchemy==3.0.5
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kafka-python==2.0.2
MarkupSafe==2.1.2
mpmath==1.2.1
multidict==6.0.4
networkx==3.0
packageurl-python==0.11.1
packaging==23.0
pandas==2.0.3
Pillow==9.4.0
pykwalify==1.8.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2023.3
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
sympy==1.11.1
typing_extensions==4.4.0
mysql-connector-python==8.1.0
tzdata==2023.3
urllib3==1.26.13
Werkzeug==2.3.6
west==1.0.0
yarl==1.9.2
